Added common component for remote user time and scheduled post indicator
