+++ webapp/channels/src/components/advanced_text_editor/advanced_text_editor.tsx
@@ -52,7 +52,6 @@ import type {PostDraft} from 'types/store/draft';
 import DoNotDisturbWarning from './do_not_disturb_warning';
 import FormattingBar from './formatting_bar';
 import {FormattingBarSpacer, Separator} from './formatting_bar/formatting_bar';
-import RemoteUserHour from './remote_user_hour';
 import SendButton from './send_button';
 import ShowFormat from './show_formatting';
 import TexteditorActions from './texteditor_actions';
@@ -67,6 +66,7 @@ import useTextboxFocus from './use_textbox_focus';
 import useUploadFiles from './use_upload_files';
 
 import './advanced_text_editor.scss';
+import PostBoxIndicator from 'components/advanced_text_editor/post_box_indicator/post_box_indicator';
 
 const FileLimitStickyBanner = makeAsyncComponent('FileLimitStickyBanner', lazy(() => import('components/file_limit_sticky_banner')));
 
@@ -120,7 +120,6 @@ const AdvancedTextEditor = ({
     const teammateId = useSelector((state: GlobalState) => getDirectChannel(state, channelId)?.teammate_id || '');
     const teammateDisplayName = useSelector((state: GlobalState) => (teammateId ? getDisplayName(state, teammateId) : ''));
     const showDndWarning = useSelector((state: GlobalState) => (teammateId ? getStatusForUserId(state, teammateId) === UserStatuses.DND : false));
-    const showRemoteUserHour = useSelector((state: GlobalState) => !showDndWarning && Boolean(getDirectChannel(state, channelId)?.teammate_id));
 
     const canPost = useSelector((state: GlobalState) => {
         const channel = getChannel(state, channelId);
@@ -560,20 +559,12 @@ const AdvancedTextEditor = ({
                 <FileLimitStickyBanner/>
             )}
             {showDndWarning && <DoNotDisturbWarning displayName={teammateDisplayName}/>}
-            <div className='postBoxIndicator'>
-                {showRemoteUserHour && (
-                    <RemoteUserHour
-                        teammateId={teammateId}
-                        displayName={teammateDisplayName}
-                    />
-                )}
-
-                <ScheduledPostIndicator
-                    location={location}
-                    channelId={channelId}
-                    postId={postId}
-                />
-            </div>
+            <PostBoxIndicator
+                channelId={channelId}
+                teammateDisplayName={teammateDisplayName}
+                location={location}
+                postId={postId}
+            />
             <div
                 className={classNames('AdvancedTextEditor', {
                     'AdvancedTextEditor__attachment-disabled': !canUploadFiles,
