+++ webapp/channels/src/components/advanced_text_editor/post_box_indicator/style.scss
@@ -1,0 +1,2 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
