+++ webapp/channels/src/components/advanced_text_editor/remote_user_hour.tsx
@@ -1,21 +1,14 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {DateTime} from 'luxon';
-import React, {useState, useEffect} from 'react';
+import React from 'react';
 import {FormattedMessage} from 'react-intl';
-import {useSelector} from 'react-redux';
 import styled from 'styled-components';
 
-import type {GlobalState} from '@mattermost/types/store';
-
-import {getTimezoneForUserProfile} from 'mattermost-redux/selectors/entities/timezone';
-import {getUser} from 'mattermost-redux/selectors/entities/users';
+import type {UserTimezone} from '@mattermost/types/users';
 
 import Moon from 'components/common/svg_images_components/moon_svg';
 import Timestamp from 'components/timestamp';
-
-import Constants from 'utils/constants';
 
 const Container = styled.div`
     display: flex;
@@ -45,48 +38,12 @@ const Icon = styled(Moon)`
 `;
 
 type Props = {
-    teammateId: string;
     displayName: string;
+    timestamp: number;
+    teammateTimezone: UserTimezone;
 }
 
-const DEFAULT_TIMEZONE = {
-    useAutomaticTimezone: true,
-    automaticTimezone: '',
-    manualTimezone: '',
-};
-
-const RemoteUserHour = ({teammateId, displayName}: Props) => {
-    const [timestamp, setTimestamp] = useState(0);
-    const [showIt, setShowIt] = useState(false);
-
-    const teammateTimezone = useSelector((state: GlobalState) => {
-        const teammate = teammateId ? getUser(state, teammateId) : undefined;
-        return teammate ? getTimezoneForUserProfile(teammate) : DEFAULT_TIMEZONE;
-    }, (a, b) => a.automaticTimezone === b.automaticTimezone &&
-        a.manualTimezone === b.manualTimezone &&
-        a.useAutomaticTimezone === b.useAutomaticTimezone);
-
-    useEffect(() => {
-        const teammateUserDate = DateTime.local().setZone(teammateTimezone.useAutomaticTimezone ? teammateTimezone.automaticTimezone : teammateTimezone.manualTimezone);
-        setTimestamp(teammateUserDate.toMillis());
-        setShowIt(teammateUserDate.get('hour') >= Constants.REMOTE_USERS_HOUR_LIMIT_END_OF_THE_DAY || teammateUserDate.get('hour') < Constants.REMOTE_USERS_HOUR_LIMIT_BEGINNING_OF_THE_DAY);
-
-        const interval = setInterval(() => {
-            const teammateUserDate = DateTime.local().setZone(teammateTimezone.useAutomaticTimezone ? teammateTimezone.automaticTimezone : teammateTimezone.manualTimezone);
-            setTimestamp(teammateUserDate.toMillis());
-            setShowIt(teammateUserDate.get('hour') >= Constants.REMOTE_USERS_HOUR_LIMIT_END_OF_THE_DAY || teammateUserDate.get('hour') < Constants.REMOTE_USERS_HOUR_LIMIT_BEGINNING_OF_THE_DAY);
-        }, 1000 * 60);
-        return () => clearInterval(interval);
-    }, [teammateTimezone.useAutomaticTimezone, teammateTimezone.automaticTimezone, teammateTimezone.manualTimezone]);
-
-    if (!showIt) {
-        return null;
-    }
-
-    if (timestamp === 0) {
-        return null;
-    }
-
+const RemoteUserHour = ({displayName, timestamp, teammateTimezone}: Props) => {
     return (
         <Container className='RemoteUserHour'>
             <Icon className='icon moonIcon'/>
